PR fixes

+++ djv-product/src/main/resources/db/migration/V2.1__create_product_table.sql
@@ -3,6 +3,6 @@ CREATE TABLE ${tables.product.name} (
   ${tables.product.columns.name} VARCHAR(255) NOT NULL,
   ${tables.product.columns.description} VARCHAR(80000),
   ${tables.product.columns.creationDate} TIMESTAMP,
-  ${tables.product.columns.category} INT,
+  ${tables.product.columns.category} BIGINT,
   ${tables.product.columns.price} DECIMAL,
 );
